+++ app/src/main/java/org/librefit/enums/healthConnect/HealthConnectStatus.kt
@@ -1,0 +1,16 @@
+/*
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ * Copyright (c) 2026. The LibreFit Contributors
+ *
+ * LibreFit is subject to additional terms covering author attribution and trademark usage;
+ * see the ADDITIONAL_TERMS.md and TRADEMARK_POLICY.md files in the project root.
+ */
+
+package org.librefit.enums.healthConnect
+
+enum class HealthConnectStatus {
+    UNAVAILABLE,
+    NEEDS_PERMISSIONS,
+    READY,
+    EXPORTING
+}
