+++ app/src/main/java/org/librefit/ui/screens/settings/SettingsScreen.kt
@@ -50,9 +50,12 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import org.librefit.R
+import org.librefit.enums.healthConnect.HealthConnectClickAction
+import org.librefit.enums.healthConnect.HealthConnectStatus
 import org.librefit.enums.userPreferences.DialogPreference
 import org.librefit.enums.userPreferences.Language
 import org.librefit.enums.userPreferences.ThemeMode
+import org.librefit.ui.models.HealthConnectState
 import org.librefit.nav.Route
 import org.librefit.ui.components.HeadlineText
 import org.librefit.ui.components.LibreFitLazyColumn
